backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.9",
    "fastapi>=0.136.1",
    "open-clip-torch>=3.3.0",
    "pillow>=12.2.0",
    "python-multipart>=0.0.28",
    "torch>=2.11.0",
    "uvicorn>=0.46.0",
]
